refactor: dto를 response로 수정
