+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/cart/CartItemResponse.kt
@@ -3,8 +3,8 @@ package woowacourse.shopping.data.remote.dto.response.cart
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CartDto(
+data class CartItemResponse(
     val id: Long,
-    val product: CartProductDto,
+    val product: CartProductResponse,
     val quantity: Int,
 )
