+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/cart/CartProductResponse.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.remote.dto.response.cart
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CartProductDto(
+data class CartProductResponse(
     val category: String,
     val id: Long,
     val imageUrl: String,
